Update transform.py
Added np.where into output of class_subset function.  This was done to be more numpythonic, in hopes of optimization.  Also, this serves a very practical purpose in the case the response vector is given in a column form, in which the previous version cannot handle. (I encountered this issue while creating the skew distance). @tamirbennatan